Create Sentiment Analysis using TextBlob
Here I've used Textblob library, passing the text in the sentiment method of library gives you 2 output values. 
1) Subjectivity : means if the text have any sentence that expresses some personal feelings, views, beliefs, opinions, allegations, desires,
beliefs, suspicions, and speculations.
2) Polarity: means if the text is positive text, negative text or neutral. it has the polarity value from 1 to -1. 
polarity close to -1 indicates the sentence is negative, polarity close to 11 indicates the sentence is positive and 0 means neutral sentence.
After that using textblob.correction method we can correct some errors is our texts. so in this meeting we have done research and work on this feature.